requirements.txt
numpy
pandas
requests
BeautifulSoup4
regex
python-dotenv
psycopg2-binary
facebook-business
flatten_json
google.cloud
datetime
sqlalchemy
numpy-financial
google-ads
soda-core
soda-core-postgres
apache-airflow-providers-http
apache-airflow-providers-airbyte==3.8.0